chore: Reorder validation steps in pipeline template (#1824)
#### Details

The current validation steps are 

##### Motivation

Make it easier to determine the successful and failing cases. The cases
are called out as succeeding or failing and are separately numbered to
make it easier to add new cases to each in the future.

The `success-3` case is currently failing due to #1713, so we will need
to update the template once that is fixed. The final result can be found
at
https://dev.azure.com/accessibility-insights-private/Accessibility%20Insights%20(private)/_build/results?buildId=47465&view=results

##### Context

<!-- Are there any parts that you've intentionally left out-of-scope for
a later PR to handle? -->
The commits here are intentionally small, which might help simplify the
review process

<!-- Were there any alternative approaches you considered? What
tradeoffs did you consider? -->

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->
- [n/a] Addresses an existing issue: 
- [n/a] Added relevant unit test for your changes. (`yarn test`)
- [n/a] Verified code coverage for the changes made. Check coverage
report at: `<rootDir>/test-results/unit/coverage`
- [x] Ran precheckin (`yarn precheckin`)