chore: Use auto-generated docs file (#1853)
#### Details

In #1820, manual edits were made directly into inputs.md file. This is a
tool-generated file, so manual edits get lost when the file is next
generated. This PR attempts to restate the core messaging of #1820, but
to do it in a way that will persist when the docs are regenerated.

##### Motivation

Follow original intent that was previously overlooked, fix `yarn
precheckin`

##### Context

<!-- Are there any parts that you've intentionally left out-of-scope for
a later PR to handle? -->

<!-- Were there any alternative approaches you considered? What
tradeoffs did you consider? -->

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->
- [n/a] Addresses an existing issue: Fixes #0000
- [n/a] Added relevant unit test for your changes. (`yarn test`)
- [n/a] Verified code coverage for the changes made. Check coverage
report at: `<rootDir>/test-results/unit/coverage`
- [x] Ran precheckin (`yarn precheckin`)