feat(oncall-vendor-transition): Increase Dependabot limit and move scheduled runs to 08:30 UTC (#1869)
* Dependabot for our repos is scheduled to run at fairly different times
that are inconvenient for some geographies. This commit moves this repo's
dependabot time to 08:30 UTC (14:00 IST, 01:30 PDT, 00:30 PST).
* Web-based projects have many dependencies, and a low Dependabot PR
limit causes us to need to run Dependabot manually multiple times to
update all dependencies. This commit increases the limit to 10 for consistency with other repos.

Part of Feature 2083093 (internal access required to view).