fix: Fix telemetry initialization error by rolling back (#1866)
#### Details

Dependabot updated the applicationinsights version while our Canary
validation pipeline was running the wrong version. The Canary pipeline
is currently failing during telemetry initialization, so the simplest
step is to roll back so we're again in a working state. This includes a
fix to prevent dependabot from suggesting updates before we're ready

##### Motivation

We should always be in a shippable state

##### Context

<!-- Are there any parts that you've intentionally left out-of-scope for
a later PR to handle? -->

<!-- Were there any alternative approaches you considered? What
tradeoffs did you consider? -->
We could fix the telemetry initialization, but we don't know how hard
that will be. This is the quick fix until we're ready for the "real" fix

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->
- [n/a] Addresses an existing issue: Fixes
- [n/a] Added relevant unit test for your changes. (`yarn test`)
- [n/a] Verified code coverage for the changes made. Check coverage
report at: `<rootDir>/test-results/unit/coverage`
- [x] Ran precheckin (`yarn precheckin`)