chore: Remove unneeded condition (#1900)
#### Details

When #1874 rearranged the validation tests, an unnecessary condition was
left in place. it's unnecessary because the run should fail if either of
the 2 previous test steps fail. Leaving the condition in place doesn't
impact the overall result (thanks to the validation job), but it's both
slower and inconsistent with the comment on lines 50 and 51.

Validation run is at
https://dev.azure.com/accessibility-insights-private/Accessibility%20Insights%20(private)/_build/results?buildId=48628&view=results

##### Motivation

Avoid maintaining unnecessary stuff

##### Context

<!-- Are there any parts that you've intentionally left out-of-scope for
a later PR to handle? -->

<!-- Were there any alternative approaches you considered? What
tradeoffs did you consider? -->

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->
- [n/a] Addresses an existing issue: Fixes
- [n/a] Added relevant unit test for your changes. (`yarn test`)
- [ ] Verified code coverage for the changes made. Check coverage report
at: `<rootDir>/test-results/unit/coverage`
- [ ] Ran precheckin (`yarn precheckin`)