feat(1ES Pipeline Migration): Migrated staging and canary ado extension validation pipelines to 1ES templates (#1963)
#### Details

Migrated publish ado-extension-canary-validation and
ado-extension-staging-validation to 1ES templates. Tested the updated
pipelines and it running as expected and have verified the artifacts
generated with old run.

Also update ado-extension-prod-validation.yaml file as it is using the
same template which modified for staging and canary validation
pipelines. However, this file is not used in any of the pipelines.

##### Motivation

This PR is part of migrating all build pipeline to 1ES template feature
work.
- [User Story
2126512](https://dev.azure.com/mseng/1ES/_workitems/edit/2126512):
[Action] Migrate ado-extension-staging-validation to 1ES PT
- [User Story
2126511](https://dev.azure.com/mseng/1ES/_workitems/edit/2126511):
[Action] Migrate ado-extension-canary-validation to 1ESPT

##### Context

<!-- Are there any parts that you've intentionally left out-of-scope for
a later PR to handle? -->

<!-- Were there any alternative approaches you considered? What
tradeoffs did you consider? -->

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->
- [na] Addresses an existing issue: Fixes #0000
- [na] Added relevant unit test for your changes. (`yarn test`)
- [na] Verified code coverage for the changes made. Check coverage
report at: `<rootDir>/test-results/unit/coverage`
- [x] Ran precheckin (`yarn precheckin`)