Breaking Change: Update node version to 20.x (#2220)
#### Details
Node 16 left support in October 2023. Updating the code node version to
node 20.x.x version. As lot of other packages are using or going to use
Node 20.x.x.

##### Motivation

Get on a supported version that we expect to last for a while

##### Context

<!-- Are there any parts that you've intentionally left out-of-scope for
a later PR to handle? -->

<!-- Were there any alternative approaches you considered? What
tradeoffs did you consider? -->

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->
- [n/a] Addresses an existing issue: Fixes #0000
- [n/a] Added relevant unit test for your changes. (`yarn test`)
- [x] Verified code coverage for the changes made. Check coverage report
at: `<rootDir>/test-results/unit/coverage`
- [x] Ran precheckin (`yarn precheckin`)